producer/requirements.txt
requests
kafka-python
websockets